first_training/ai-agent-training/pyproject.toml
[project]
name = "ai-agent-training"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11.9"
dependencies = [
    "langchain==0.1.17",
    "openai>=2.3.0",
    "python-dotenv>=1.1.1",
]
